add javadoc copyright, fix package, format code style
