+++ dubbo/src/main/java/com/alibaba/fescar/samples/dubbo/ApplicationKeeper.java
@@ -16,12 +16,12 @@
 
 package com.alibaba.fescar.samples.dubbo;
 
+import java.util.concurrent.locks.Condition;
+import java.util.concurrent.locks.ReentrantLock;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.support.AbstractApplicationContext;
-
-import java.util.concurrent.locks.Condition;
-import java.util.concurrent.locks.ReentrantLock;
 
 /**
  * The type Application keeper.
