--- springboot/src/main/java/com/alibaba/fescar/demo/dubbo/AssetService.java
@@ -1,6 +1,0 @@
-package com.alibaba.fescar.demo.dubbo;
-
-public interface AssetService {
-
-	public int increase();
-}
