--- springboot/src/main/java/com/alibaba/fescar/demo/service/AssignService.java
@@ -1,7 +1,0 @@
-package com.alibaba.fescar.demo.service;
-
-import com.alibaba.fescar.demo.sys.domain.AssetAssign;
-
-public interface AssignService {
-	AssetAssign increaseAmount(String id);
-}
