--- springboot/src/main/java/com/alibaba/fescar/demo/sys/domain/Asset.java
@@ -1,49 +1,0 @@
-package com.alibaba.fescar.demo.sys.domain;
-
-import javax.persistence.*;
-import java.io.Serializable;
-import java.math.BigDecimal;
-
-/**
- * The persistent class for the t_asset database table.
- * 
- */
-@Entity
-@Table(name = "t_asset")
-@NamedQuery(name = "Asset.findAll", query = "SELECT a FROM Asset a")
-public class Asset implements Serializable {
-	private static final long serialVersionUID = 1L;
-
-	@Id
-	private String id;
-
-	private BigDecimal amount;
-
-	@Column(name = "voucher_code")
-	private String voucherCode;
-
-	public String getId() {
-		return id;
-	}
-
-	public void setId(String id) {
-		this.id = id;
-	}
-
-	public BigDecimal getAmount() {
-		return amount;
-	}
-
-	public void setAmount(BigDecimal amount) {
-		this.amount = amount;
-	}
-
-	public String getVoucherCode() {
-		return voucherCode;
-	}
-
-	public void setVoucherCode(String voucherCode) {
-		this.voucherCode = voucherCode;
-	}
-
-}
